vika palautus

+++ app/controllers/category_controller.php
@@ -8,10 +8,15 @@
 			$attributes = array(
 				'name' => $params['name']
 			);
-
 			$category = new Category($attributes);
-			$category->save();
-			Redirect::to('/task', array('message' => 'Luokka lisätty!'));
+			$errors = $category->errors();
+			if (count($errors) == 0) {
+				$category->save();
+				Redirect::to('/task', array('message' => 'Luokka lisätty!'));				
+			} else {
+				View::make('category/new.html', array('errors' => $errors));
+			}
+
 		}
 
 		public static function index() {
@@ -21,7 +26,34 @@
 
 		public static function destroy($id) {
 			$category = new Category(array('id' => $id));
-			$category->destroy();
-			Redirect::to('/category');
+			if ($category->voikoPoistaa()) {
+				$category->destroy();
+				Redirect::to('/category');				
+			} else {
+				Redirect::to('/category', array('message' => 'Luokkaa ei voi poistaa jos se on käytössä jollain askareella!'));
+			}
+
+		}
+
+		public static function edit($id) {
+			$category = Category::find($id);
+			View::make('category/edit.html', array('attributes' => $category));
+		}
+
+		public static function update($id) {
+			$params = $_POST;
+			$attributes = array(
+				'id' => $id,
+				'name' => $params['name']
+			);
+
+			$category = new Category($attributes);
+			$errors = $category->errors();
+			if (count($errors) > 0) {
+				View::make('category/edit.html', array('errors' => $errors, 'attributes' => $category));	
+			} else {
+				$category->update();
+				Redirect::to('/category', array('message' => 'Luokkaa muokattu onnistuneesti!'));
+			}
 		}
 	}
